Add files via upload
Titanic was a British passenger liner that sank in the North Atlantic Ocean in 1912 after the ship struck an iceberg during her maiden voyage from Southampton to New York City. Titanic data is dataset containing details of passengers travelled on this ship.
Objective of analysis is to use the data visualisation packages for understanding of data.